+++ simconnect-sdk/src/lib.rs
@@ -4,7 +4,7 @@
 //! ## Usage
 //! ```toml
 //! [dependencies]
-//! simconnect-sdk = { version = "0.1", features = ["derive"] }
+//! simconnect-sdk = { version = "0.2", features = ["derive"] }
 //! ```
 //!
 //! ```rust,no_run