+++ CHANGELOG.md
@@ -6,6 +6,8 @@ file. This change log follows the conventions of
 
 ## [Unreleased]
 
+## [v0.2.0] - 2022-10-29
+
 ### Added
 
 - `Notification::SystemEvent`, `SystemEventRequest` and `SystemEvent` have been added. System Events can be subscribed to by using `SimConnect::subscribe_to_system_event` and unsubscribed from by using `SimConnect::unsubscribe_from_system_event`.
@@ -50,6 +52,7 @@ file. This change log follows the conventions of
 ### Initial Release of simconnect-sdk
 
 [unreleased]: https://github.com/mihai-dinculescu/simconnect-sdk
+[v0.2.0]: https://github.com/mihai-dinculescu/simconnect-sdk/tree/v0.2.0
 [v0.1.3]: https://github.com/mihai-dinculescu/simconnect-sdk/tree/v0.1.3
 [v0.1.2]: https://github.com/mihai-dinculescu/simconnect-sdk/tree/v0.1.2
 [v0.1.1]: https://github.com/mihai-dinculescu/simconnect-sdk/tree/v0.1.1